Move supporting functions to utils.unpacking.
